chore(core): move blocks api to its own file
